requirements.txt
tensorflow==2.15.0.post1
pandas
dvc
mlflow==2.10.2
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-pyYAML
scipy
Flask
Flask-Cors
gdown
python-dotenv
-e .